♻️ refactor: Extract functions from main.js into util and component folders
- Moved individual functions from main.js to separate files for better modularity
- Placed utility functions inside the util folder
- Moved component creation functions into the component folder
- Improved code organization and maintainability
